refactor: conan cli: fix options pkg and name
